backend/requirements.txt
attrs==23.2.0
django==5.0.7
djangorestframework==3.15.2
celery==5.4.0
mysqlclient==2.2.4
python-dotenv==1.0.1
oauthlib==3.2.2
pyjwt==2.8.0